requirements.txt
astroid==2.3.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.1.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
mtnmomo==3.0.1
phonenumbers==8.10.20
pylint==2.4.2
python-dotenv==0.10.3
requests==2.21.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.24.3
Werkzeug==0.16.0
wrapt==1.11.2
